// src/common/interfaces/category.interfaces.ts
import { Types } from "mongoose";
import { IUser } from "./user.interfaces";
import { IBrand } from "./brand.interfaces";

export interface ICategory {
    _id?: Types.ObjectId,

    name: string,
    slug: string,
    description?: string
    image: string,
    assetFolderId: string,

    brands?: Types.ObjectId[] | IBrand[]
    createdBy: Types.ObjectId | IUser
    updatedBy?: Types.ObjectId | IUser


    createdAt?: Date;
    updateAt?: Date;

    freezedAt?: Date | undefined;
    restoredAt?: Date | undefined;




}